refactor: reorganize ReportSummary components
