+++ packages/web-reporter-ui/src/sections/ReportSummary/OldReportSummaryComponents.tsx
@@ -1,20 +1,19 @@
-import React, { FunctionComponent } from "react";
-import { AveragedTestCaseResult, TestCaseResult } from "@perf-profiler/types";
+import { Tooltip, Button } from "@mui/material";
 import {
-  getAverageCpuUsage,
-  getAverageFPSUsage,
-  getAverageRAMUsage,
-  getAverageTotalHightCPUUsage,
   sanitizeProcessName,
+  getAverageFPSUsage,
+  getAverageCpuUsage,
+  getAverageRAMUsage,
 } from "@perf-profiler/reporter";
-import { roundToDecimal } from "../../utils/roundToDecimal";
-import { Score } from "../components/Score";
-import { orderBy } from "lodash";
-import Button from "@mui/material/Button";
+import { TestCaseResult, AveragedTestCaseResult } from "@perf-profiler/types";
 import FileDownloadIcon from "@mui/icons-material/FileDownload";
-import Tooltip from "@mui/material/Tooltip";
-import { exportRawDataToJSON } from "../../utils/reportRawDataExport";
-import { MetricWithExplanation } from "../components/ReportSummary/MetricWithExplanation";
+import { orderBy } from "lodash";
+import React from "react";
+import { exportRawDataToJSON } from "../../../utils/reportRawDataExport";
+import { roundToDecimal } from "../../../utils/roundToDecimal";
+import { MetricWithExplanation } from "../../components/ReportSummary/MetricWithExplanation";
+import { Score } from "../../components/Score";
+import { SimpleTable } from "../../components/SimpleTable";
 
 const HighCpuProcesses = ({
   highCpuProcesses,
@@ -54,7 +53,7 @@ const HighCpuProcesses = ({
   );
 };
 
-export const ReportSummary = ({
+export const OldReportSummary = ({
   results,
   averagedResults,
 }: {
@@ -188,104 +187,5 @@ export const ReportSummary = ({
     ],
   ];
 
-  return (
-    <div className="flex flex-row overflow-x-scroll px-32 pt-12">
-      {averagedResults
-        .map((result) => (
-          <ReportSummaryCard key={result.name} averagedResult={result} />
-        ))
-        .reduce((acc, curr, index) => {
-          if (index === 0) return [curr];
-          return [...acc, <div className="w-24" />, curr];
-        }, [] as React.ReactNode[])}
-    </div>
-  );
+  return <SimpleTable rows={table} />;
 };
-
-type ReportSummaryCardProps = {
-  averagedResult: AveragedTestCaseResult;
-};
-
-const ReportSummaryCard: FunctionComponent<ReportSummaryCardProps> = ({
-  averagedResult,
-}) => {
-  return (
-    <div className="flex flex-col items-center py-6 px-10 bg-dark-charcoal border border-gray-800 rounded-lg w-[520px]">
-      <div className="text-neutral-300">{averagedResult.name}</div>
-
-      <div className="h-8" />
-
-      <Score result={averagedResult} />
-
-      <div className="h-8" />
-
-      <ReportSummaryCardInfoRow
-        title="Average Test Runtime"
-        value={roundToDecimal(
-          getAverageCpuUsage(averagedResult.average.measures),
-          1
-        )}
-        unit="ms"
-      />
-      <div className="h-2" />
-      <ReportSummaryCardInfoRow
-        title="Average FPS"
-        value={roundToDecimal(
-          getAverageRAMUsage(averagedResult.average.measures),
-          1
-        )}
-        unit="FPS"
-      />
-      <div className="h-2" />
-      <ReportSummaryCardInfoRow
-        title="Average CPU usage"
-        value={roundToDecimal(
-          getAverageCpuUsage(averagedResult.average.measures),
-          1
-        )}
-        unit="%"
-      />
-      <div className="h-2" />
-      <ReportSummaryCardInfoRow
-        title="Hight CPU Usage"
-        value={roundToDecimal(
-          getAverageTotalHightCPUUsage(averagedResult.averageHighCpuUsage) /
-            1000,
-          1
-        )}
-        unit="s"
-      />
-      <div className="h-2" />
-      <ReportSummaryCardInfoRow
-        title="Average RAM usage"
-        value={roundToDecimal(
-          getAverageRAMUsage(averagedResult.average.measures),
-          1
-        )}
-        unit="MB"
-      />
-    </div>
-  );
-};
-
-type ReportSummaryCardInfoRowProps = {
-  title: string;
-  value: number;
-  unit?: string;
-};
-
-const ReportSummaryCardInfoRow: FunctionComponent<
-  ReportSummaryCardInfoRowProps
-> = ({ title, value, unit }) => (
-  <div className="flex flex-row w-full px-6 py-4 border rounded-lg border-gray-800">
-    <div className="text-neutral-300">{title}</div>
-
-    <div className="grow" />
-
-    <p className="text-neutral-300 whitespace-pre">{`${value} ${unit}`}</p>
-
-    <div className="w-2" />
-
-    <div className="text-neutral-300"> v </div>
-  </div>
-);
